Try to fix linking with libraries in /app/lib64
When building a flatpak module, make sure to pass -L%{_libdir} to ld so
that it can find libraries installed in /app/lib64. Normally, this is
set through the %{__global_ldflags} macro, but firefox doesn't use that,
so we need to pass the correct -L manually.

This should hopefully fix the issue with pipewire-0.2 linking that
recent firefox flatpak builds have run into.