Make provides/requires filtering smarter/automatic
The issue with the previous approach was that _everything_ that
firefox's shared libraries required was filtered out. This included e.g.
libxul.so's dependency on libdbus-glib-1.so.2 which definitely should
not be filtered out.

Instead, a better approach is to only filter out all of the private
libraries provides (as they are firefox's implementation detail) and
also all the internal requires on the same private libraries.

RPM doesn't make it very easy, but the rudimentary filtering is there
and we can build upon that. This commit improves the filtering by making
it actually look at what's built, and automatically generating the list
of private libraries to filter based on that.

This also lets us drop manual dbus-glib requires and instead rely on
rpm's automatic dep generator (the results are no longer filtered out).

This should make the filtering future proof when more private libraries
are added in the future (e.g. nss).

https://bugzilla.redhat.com/show_bug.cgi?id=1582116