Remove gtk2 support as flash plugin is no longer supported
ESR91 is supposed to drop gtk2 support, but in the mean time, we can
just rm the gtk2 libmozgtk.so as we are using gtk3 by default. The last
supported plugin -- adobe flash -- is no longer supported in Firefox 85,
so we no longer need to keep the gtk2 libmozgtk around for the plugin
loader.

https://pagure.io/fedora-workstation/issue/180