Add -g1 to find-debuginfo flags to try and get builds through
F38 and F39 builds are failing a lot on x86_64 with gdb-add-index
being killed, even with the %limit_build thing I added previously,
so we'll have to try this. This is a big hammer which reduces
the quality of backtraces a lot, but we really need to get
builds through right now to fix major bugs.